+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/LeagueArchive/BlueLeft_Fast.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.LeagueArchive;
 
 import com.acmerobotics.dashboard.FtcDashboard;
 import com.acmerobotics.dashboard.config.Config;
@@ -7,6 +7,7 @@ import com.acmerobotics.roadrunner.geometry.Pose2d;
 import com.acmerobotics.roadrunner.geometry.Vector2d;
 import com.qualcomm.hardware.dfrobot.HuskyLens;
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.Servo;
@@ -17,15 +18,14 @@ import org.firstinspires.ftc.teamcode.drive.DriveConstants;
 import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 import org.firstinspires.ftc.teamcode.trajectorysequence.TrajectorySequence;
 
-import java.util.Vector;
-
 
 /*
  * This is a simple routine to test translational drive capabilities.
  */
 @Config
-@Autonomous(name = "RedRight_Pixel", group = "4RedRight")
-public class RedRight_Pixel extends LinearOpMode {
+@Disabled
+@Autonomous(name = "BlueLeft_Fast", group = "1BlueLeft")
+public class BlueLeft_Fast extends LinearOpMode {
     // initialize motors
     private DcMotor backLeft;
     private DcMotor frontLeft;
@@ -33,7 +33,6 @@ public class RedRight_Pixel extends LinearOpMode {
     private DcMotor backRight;
 
     private DcMotor armMotor;
-
     private Servo grabberServo;
     private Servo wristServo;
 
@@ -59,7 +58,6 @@ public class RedRight_Pixel extends LinearOpMode {
         armMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         armMotor.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
-        armMotor.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
         //define huskylens
         huskyLens = hardwareMap.get(HuskyLens.class, "huskylens");
@@ -82,31 +80,28 @@ public class RedRight_Pixel extends LinearOpMode {
 
         int speed = 25; // inches per second
         int slow = 20;
-        int armHeight = -5025; //TODO used to be -5150
+        int armHeight = -5050; //TODO used to be -5150
 
         Pose2d startPose = new Pose2d(0,0,0);
 
         drive.setPoseEstimate(startPose);
 
         // TODO change forward and back to lines so it uses the PID
-
 //spikeLeft
         TrajectorySequence spikeLeft = drive.trajectorySequenceBuilder()
-                .lineToLinearHeading(
-                        new Pose2d(19, -5, Math.toRadians(60)),
+                .splineToConstantHeading(
+                        new Vector2d(21.5, 10.5), 0,
                         SampleMecanumDrive.getVelocityConstraint(speed, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
                         SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
                 )
-                .forward(9,
-                        SampleMecanumDrive.getVelocityConstraint(speed, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
-                        SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
-                )
-                .back(9,
+                //TODO change to linearHeading
+                .lineToConstantHeading(
+                        new Vector2d(16.5, 10.5),
                         SampleMecanumDrive.getVelocityConstraint(speed, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
                         SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
                 )
                 .lineToLinearHeading(
-                        new Pose2d(30.8, -33.3, Math.toRadians(-90)),
+                        new Pose2d(18, 32.5, Math.toRadians(90)),
                         SampleMecanumDrive.getVelocityConstraint(speed, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
                         SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
                 )
@@ -123,7 +118,7 @@ public class RedRight_Pixel extends LinearOpMode {
 //spikeLeftPark
         TrajectorySequence spikeLeftPark = drive.trajectorySequenceBuilder(spikeLeftBackUp.end())
                 .lineToConstantHeading(
-                        new Vector2d(4, -33),
+                        new Vector2d(4, 32.5),
                         SampleMecanumDrive.getVelocityConstraint(slow, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
                         SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
                 )
@@ -142,7 +137,7 @@ public class RedRight_Pixel extends LinearOpMode {
                         SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
                 )
                 .lineToLinearHeading(
-                        new Pose2d(26, -33.3, Math.toRadians(-90)),
+                        new Pose2d(25, 32.5, Math.toRadians(90)),
                         SampleMecanumDrive.getVelocityConstraint(speed, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
                         SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
                 )
@@ -150,52 +145,38 @@ public class RedRight_Pixel extends LinearOpMode {
 
 //spikeCenterBackUp
         TrajectorySequence spikeCenterBackUp = drive.trajectorySequenceBuilder(spikeCenter.end())
-                .back(3,
+                .back(2,
                         SampleMecanumDrive.getVelocityConstraint(10, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
                         SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
                 )
                 .build();
 
-//spikeCenterPixel
-        TrajectorySequence spikeCenterPixel = drive.trajectorySequenceBuilder(spikeCenterBackUp.end())
-                .lineToLinearHeading(
-                        new Pose2d(29.25, 60, Math.toRadians(-90)),
-                        SampleMecanumDrive.getVelocityConstraint(45, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
+//spikeCenterPark
+        TrajectorySequence spikeCenterPark = drive.trajectorySequenceBuilder(spikeCenterBackUp.end())
+                .lineToConstantHeading(
+                        new Vector2d(4, 32.5),
+                        SampleMecanumDrive.getVelocityConstraint(slow, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
                         SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
                 )
-                .build();
-
-//spikeCenterPixelPickUp
-        TrajectorySequence spikeCenterPixelPickUp = drive.trajectorySequenceBuilder(spikeCenterPixel.end())
-                .lineToLinearHeading(
-                        new Pose2d(29, 71.5, Math.toRadians(-90)),
-                        SampleMecanumDrive.getVelocityConstraint(10, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
-                        SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
-                )
-                .build();
-
-//spikeCenterPixelPark
-        TrajectorySequence spikeCenterPixelPark = drive.trajectorySequenceBuilder(spikeCenterPixelPickUp.end())
-                .lineToLinearHeading(
-                        new Pose2d(29, -31, Math.toRadians(-90))
-                )
-                .turn(Math.toRadians(180))
                 .build();
 
 //spikeRight
         TrajectorySequence spikeRight = drive.trajectorySequenceBuilder()
-                .splineToConstantHeading(
-                        new Vector2d(21.5, -10.5), 0,
-                        SampleMecanumDrive.getVelocityConstraint(40, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
+                .lineToLinearHeading(
+                        new Pose2d(19, 5, Math.toRadians(-60)),
+                        SampleMecanumDrive.getVelocityConstraint(speed, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
                         SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
                 )
-                .lineToConstantHeading(
-                        new Vector2d(16.5, -10.5),
-                        SampleMecanumDrive.getVelocityConstraint(40, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
+                .forward(9,
+                        SampleMecanumDrive.getVelocityConstraint(speed, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
+                        SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
+                )
+                .back(9,
+                        SampleMecanumDrive.getVelocityConstraint(speed, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
                         SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
                 )
                 .lineToLinearHeading(
-                        new Pose2d(17.7, -33.3, Math.toRadians(-90)),
+                        new Pose2d(33, 32.5, Math.toRadians(90)),
                         SampleMecanumDrive.getVelocityConstraint(speed, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
                         SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
                 )
@@ -209,56 +190,15 @@ public class RedRight_Pixel extends LinearOpMode {
                 )
                 .build();
 
-//spikeRightPixel
-        TrajectorySequence spikeRightPixel = drive.trajectorySequenceBuilder(spikeRightBackUp.end())
-                .lineToLinearHeading(
-                        new Pose2d(4, 0, Math.toRadians(-90)),
-                        SampleMecanumDrive.getVelocityConstraint(35, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
-                        SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
-                )
-                .lineToLinearHeading(
-                        new Pose2d(4, 50, Math.toRadians(-90))
-//                        ,
-//                        SampleMecanumDrive.getVelocityConstraint(45, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
-//                        SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
-                )
-                .lineToLinearHeading(
-                        new Pose2d(30.5, 62.5, Math.toRadians(-90)),
-                        SampleMecanumDrive.getVelocityConstraint(35, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
-                        SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
-                )
-
-                .build();
-
-//spikeRightPixelPickUp
-        TrajectorySequence spikeRightPixelPickUp = drive.trajectorySequenceBuilder(spikeRightPixel.end())
-                .lineToLinearHeading(
-                        new Pose2d(30.5, 71.5, Math.toRadians(-90)),
-                        SampleMecanumDrive.getVelocityConstraint(10, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
+//spikeRightPark
+        TrajectorySequence spikeRightPark = drive.trajectorySequenceBuilder(spikeRightBackUp.end())
+                .lineToConstantHeading(
+                        new Vector2d(4, 32.5),
+                        SampleMecanumDrive.getVelocityConstraint(slow, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
                         SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
                 )
                 .build();
 
-//spikeRightPixelPark
-        TrajectorySequence spikeRightPixelPark = drive.trajectorySequenceBuilder(spikeRightPixelPickUp.end())
-                .lineToLinearHeading(
-                        new Pose2d(30, 60, Math.toRadians(-90)),
-                        SampleMecanumDrive.getVelocityConstraint(35, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
-                        SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
-                )
-                .build();
-
-//spikeRightPixelPark_2
-        TrajectorySequence spikeRightPixelPark_2 = drive.trajectorySequenceBuilder(spikeRightPixelPark.end())
-                .lineToLinearHeading(
-                        new Pose2d(5, 50, Math.toRadians(-90)),
-                        SampleMecanumDrive.getVelocityConstraint(35, DriveConstants.MAX_ANG_VEL, DriveConstants.TRACK_WIDTH),
-                        SampleMecanumDrive.getAccelerationConstraint(DriveConstants.MAX_ACCEL)
-                )
-                .lineToLinearHeading(
-                        new Pose2d(5, -30, Math.toRadians(-90))
-                )
-                .build();
 
         wristServo.setPosition(1);
         sleep(500);
@@ -287,7 +227,7 @@ public class RedRight_Pixel extends LinearOpMode {
 
                     for (int i = 0; i < blocks.length; i++) {
 
-                        if (blocks[i].id == 1 && blocks[i].width > 15) {
+                        if (blocks[i].id == 2 && blocks[i].width > 15) {
                             //TODO id #
                             x_pos = blocks[i].x;
                             telemetry.addData("x_pos: ", x_pos);
@@ -319,8 +259,8 @@ public class RedRight_Pixel extends LinearOpMode {
                 telemetry.update();
             }
             armMotor.setPower(0);
+            sleep(500);
 
-            sleep(250);
             grabberServo.setPosition(0.5);
             sleep(500);
 
@@ -328,7 +268,7 @@ public class RedRight_Pixel extends LinearOpMode {
             drive.followTrajectorySequence(spikeLeftBackUp);
 
             //arm down
-            while(armMotor.getCurrentPosition() <= -500) {
+            while(armMotor.getCurrentPosition() <= -750) {
                 armMotor.setPower(1);
                 telemetry.addData("position: ", armMotor.getCurrentPosition());
                 telemetry.update();
@@ -348,7 +288,7 @@ public class RedRight_Pixel extends LinearOpMode {
             drive.followTrajectorySequence(spikeCenter);
 
             grabberServo.setPosition(0.2);
-            sleep(250);
+            sleep(500);
 
             //arm up
             while(armMotor.getCurrentPosition() >= armHeight) {
@@ -357,8 +297,8 @@ public class RedRight_Pixel extends LinearOpMode {
                 telemetry.update();
             }
             armMotor.setPower(0);
+            sleep(500);
 
-            sleep(250);
             grabberServo.setPosition(0.5);
             sleep(500);
 
@@ -366,54 +306,15 @@ public class RedRight_Pixel extends LinearOpMode {
             drive.followTrajectorySequence(spikeCenterBackUp);
 
             //arm down
-            while(armMotor.getCurrentPosition() <= -500) {
+            while(armMotor.getCurrentPosition() <= -750) {
                 armMotor.setPower(1);
                 telemetry.addData("position: ", armMotor.getCurrentPosition());
                 telemetry.update();
             }
             armMotor.setPower(0);
 
-            //pixel
-            drive.followTrajectorySequence(spikeCenterPixel);
-
-            wristServo.setPosition(0);
-            sleep(250);
-
-            while(armMotor.getCurrentPosition() >= -625) {
-                armMotor.setPower(-0.5);
-                telemetry.addData("position: ", armMotor.getCurrentPosition());
-                telemetry.update();
-            }
-            armMotor.setPower(0);
-
-            drive.followTrajectorySequence(spikeCenterPixelPickUp);
-
-            while(armMotor.getCurrentPosition() <= -150) {
-                armMotor.setPower(1);
-                telemetry.addData("position: ", armMotor.getCurrentPosition());
-                telemetry.update();
-            }
-            armMotor.setPower(0);
-            sleep(500);
-
-            grabberServo.setPosition(0.2);
-            sleep(750);
-
-            while(armMotor.getCurrentPosition() >= -400) {
-                armMotor.setPower(-0.5);
-                telemetry.addData("position: ", armMotor.getCurrentPosition());
-                telemetry.update();
-            }
-            armMotor.setPower(0);
-            sleep(150);
-
-            drive.followTrajectorySequence(spikeCenterPixelPark);
-
-            wristServo.setPosition(0.5);
-            sleep(150);
-
-            grabberServo.setPosition(0.5);
-
+            //park
+            drive.followTrajectorySequence(spikeCenterPark);
         }
 
         if (x_pos > 213) {
@@ -425,74 +326,37 @@ public class RedRight_Pixel extends LinearOpMode {
             drive.followTrajectorySequence(spikeRight);
 
             grabberServo.setPosition(0.2);
-            sleep(250);
+            sleep(500);
 
             //arm up
             while(armMotor.getCurrentPosition() >= armHeight) {
-                armMotor.setPower(-0.6);
+                armMotor.setPower(-0.5);
                 telemetry.addData("position: ", armMotor.getCurrentPosition());
                 telemetry.update();
             }
             armMotor.setPower(0);
+            sleep(500);
 
-            sleep(250);
             grabberServo.setPosition(0.5);
-            sleep(250);
+            sleep(500);
 
             //back up
             drive.followTrajectorySequence(spikeRightBackUp);
 
             //arm down
-            while(armMotor.getCurrentPosition() <= -500) {
+            while(armMotor.getCurrentPosition() <= -750) {
                 armMotor.setPower(1);
                 telemetry.addData("position: ", armMotor.getCurrentPosition());
                 telemetry.update();
             }
             armMotor.setPower(0);
 
-            //pixel
-            drive.followTrajectorySequence(spikeRightPixel);
-
-            wristServo.setPosition(0);
-            sleep(250);
-
-            while(armMotor.getCurrentPosition() >= -625) {
-                armMotor.setPower(-1);
-                telemetry.addData("position: ", armMotor.getCurrentPosition());
-                telemetry.update();
-            }
-            armMotor.setPower(0);
-
-            drive.followTrajectorySequence(spikeRightPixelPickUp);
-
-            while(armMotor.getCurrentPosition() <= -150) {
-                armMotor.setPower(1);
-                telemetry.addData("position: ", armMotor.getCurrentPosition());
-                telemetry.update();
-            }
-            armMotor.setPower(0);
-            sleep(500);
-
-            grabberServo.setPosition(0.2);
-            sleep(500);
-
-            while(armMotor.getCurrentPosition() >= -400) {
-                armMotor.setPower(-1);
-                telemetry.addData("position: ", armMotor.getCurrentPosition());
-                telemetry.update();
-            }
-            armMotor.setPower(0);
-            sleep(150);
-
-            drive.followTrajectorySequence(spikeRightPixelPark);
-
-            wristServo.setPosition(1);
-            sleep(75);
-
-            drive.followTrajectorySequence(spikeRightPixelPark_2);
+            //park
+            drive.followTrajectorySequence(spikeRightPark);
         }
 
         //if nothing is detected go to the center spike
+        //TODO changed to left for Marquette
         if (x_pos == -1) {
             spikeZone = "left";
             telemetry.addData("spikeZone: ", spikeZone);
@@ -511,8 +375,8 @@ public class RedRight_Pixel extends LinearOpMode {
                 telemetry.update();
             }
             armMotor.setPower(0);
+            sleep(500);
 
-            sleep(250);
             grabberServo.setPosition(0.5);
             sleep(500);
 
@@ -545,13 +409,13 @@ public class RedRight_Pixel extends LinearOpMode {
 
         //get ID # for tag to detect
         if (spikeZone == "left") {
-            tagID = 4;
+            tagID = 1;
         }
         if (spikeZone == "center") {
-            tagID = 5;
+            tagID = 2;
         }
         if (spikeZone == "right") {
-            tagID = 6;
+            tagID = 3;
         }
 
         //trajectories for moving left and right
